Revert "Merge pull request #353 from vtex-apps/chore/improve-search-context"
This reverts commit 45b0268ad80591623f99547585a0e743d7527122, reversing
changes made to 72125efc9b1a5adabd36d7b19a9910796bf12f98.
