+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
-## [2.88.4] - 2019-07-02
+## [2.88.4] - 2019-07-02 [YANKED]
 
 ### Changed
